# Scanner#checkToken(Token.ID) avoids varargs