Calls to `Scanner#checkToken(Token.ID...)` with single arguments incur expensive varargs `Object[]` array allocations on very hot parsing code paths (e.g. `org.yaml.snakeyaml.parser.ParserImpl`). 

These can be avoided in JDK 8+ by adding a new `checkToken(Token.ID)` to the `Scanner` interface with default implementation, and overriding in `ScannerImpl` to provide a more efficient implementation.