@asomov 2024-08-05
Thank you for the contribution. The migration to Java 8 should happen very soon.