@schlosna 2024-08-05
> Thank you for the contribution. The migration to Java 8 should happen very soon.

Thanks for the quick review, I will follow up with the JMH benchmarks, and hold off until the JDK 8 source/target changes land in origin.