@schlosna 2024-08-06
Added JMH `ParseBenchmark` 05ea2fae running locally on `2021 Apple MacBookPro M1 Pro` with JDK 17 & JDK 21 runtimes demonstrating ~20% to ~40% improvements on small (18.34 KiB) to medium (2.17 MiB) sized yaml documents.

### VM version: JDK 17.0.11, OpenJDK 64-Bit Server VM, 17.0.11+9-LTS

## Before
```
Benchmark             (entries)  Mode  Cnt    Score     Error  Units
ParseBenchmark.load        1000  avgt    3    1.605 ±   1.170  ms/op
ParseBenchmark.load      100000  avgt    3  247.998 ±  92.043  ms/op
ParseBenchmark.parse       1000  avgt    3    1.092 ±   0.296  ms/op
ParseBenchmark.parse     100000  avgt    3  114.479 ± 177.379  ms/op
```

## After
```
Benchmark             (entries)  Mode  Cnt    Score    Error  Units
ParseBenchmark.load        1000  avgt    3    0.925 ±  0.196  ms/op
ParseBenchmark.load      100000  avgt    3  149.746 ± 61.340  ms/op
ParseBenchmark.parse       1000  avgt    3    0.602 ±  0.020  ms/op
ParseBenchmark.parse     100000  avgt    3   57.208 ±  2.728  ms/op
```


### VM version: JDK 21.0.3, OpenJDK 64-Bit Server VM, 21.0.3+9-LTS

## Before
```
Benchmark             (entries)  Mode  Cnt    Score    Error  Units
ParseBenchmark.load        1000  avgt    3    1.193 ±  0.063  ms/op
ParseBenchmark.load      100000  avgt    3  188.215 ± 13.995  ms/op
ParseBenchmark.parse       1000  avgt    3    0.895 ±  0.025  ms/op
ParseBenchmark.parse     100000  avgt    3   92.312 ±  9.998  ms/op
```

## After
```
Benchmark             (entries)  Mode  Cnt    Score    Error  Units
ParseBenchmark.load        1000  avgt    3    1.056 ±  0.074  ms/op
ParseBenchmark.load      100000  avgt    3  151.990 ± 11.969  ms/op
ParseBenchmark.parse       1000  avgt    3    0.636 ±  0.061  ms/op
ParseBenchmark.parse     100000  avgt    3   64.557 ±  1.321  ms/op
```