@schlosna 2024-08-06
Looking through JFR snapshots from the JMH benchmark show the `org.yaml.snakeyaml.tokens.Token$ID[]` vararg allocations from some of the single argument call sites.

![image](https://github.com/user-attachments/assets/89938684-4577-48eb-873f-b52076779924)