@asomov 2024-08-31
@schlosna feel free to pull the latest commits with Java 8