@asomov 2024-09-01
thank you!
https://bitbucket.org/snakeyaml/snakeyaml/wiki/Changes